[IDLE-300] 인앱 공고, 크롤링 공고 상속관계 추가
