+++ core/domain/src/main/kotlin/com/idle/domain/model/jobposting/CrawlingJobPosting.kt
@@ -1,22 +1,16 @@
 package com.idle.domain.model.jobposting
 
 import java.time.LocalDate
-import java.time.ZoneId
 
 data class CrawlingJobPosting(
-    val id: String,
+    override val id: String,
+    override val distance: Int,
+    override val applyDeadline: LocalDate,
+    override val jobPostingType: JobPostingType,
+    override val isFavorite: Boolean,
     val title: String,
-    val distance: String,
     val workingTime: String,
     val workingSchedule: String,
     val payInfo: String,
-    val applyDeadline: LocalDate,
-    val jobPostingType: JobPostingType,
     val createdAt: LocalDate,
-) {
-    fun calculateDeadline(): Long {
-        val seoulZone = ZoneId.of("Asia/Seoul")
-        val nowDate = LocalDate.now(seoulZone)
-        return applyDeadline.toEpochDay() - nowDate.toEpochDay()
-    }
-}
+) : JobPosting(id, distance, applyDeadline, jobPostingType, isFavorite)
