+++ core/domain/src/main/kotlin/com/idle/domain/model/jobposting/CrawlingJobPostingDetail.kt
@@ -4,7 +4,11 @@ import java.time.LocalDate
 import java.time.ZoneId
 
 data class CrawlingJobPostingDetail(
-    val id: String,
+    override val id: String,
+    override val distance: Int,
+    override val applyDeadline: LocalDate,
+    override val jobPostingType: JobPostingType,
+    override val isFavorite: Boolean,
     val title: String,
     val content: String,
     val clientAddress: String,
@@ -14,19 +18,10 @@ data class CrawlingJobPostingDetail(
     val payInfo: String,
     val workingTime: String,
     val workingSchedule: String,
-    val applyDeadline: LocalDate,
     val recruitmentProcess: String,
     val applyMethod: String,
     val requireDocumentation: String,
     val centerName: String,
     val centerAddress: String,
     val jobPostingUrl: String,
-    val jobPostingType: JobPostingType,
-    val isFavorite: Boolean = false,
-) {
-    fun calculateDeadline(): Long {
-        val seoulZone = ZoneId.of("Asia/Seoul")
-        val nowDate = LocalDate.now(seoulZone)
-        return applyDeadline.toEpochDay() - nowDate.toEpochDay()
-    }
-}
+) : JobPosting(id, distance, applyDeadline, jobPostingType, isFavorite)
